significant refactor of auto_chess game for clarity and extensibility
notably, instead of Card methods taking a Game as an argument (which
required some gross bookkeeping to track the active and non-active
players), we now have a GameState namedtuple/struct which provides
handles on all the information a Card needs access to. Now, giving
Card methods access to a new datum requires only adding a field to
that struct and to the Game method which constructs it.